refactor: assignment 1 from spark.sql to pyspark
