+++ assignment_1/README.md
@@ -1,55 +1,56 @@
 
 # Assignment 1 - Diário de Bordo
 
-## 📌 Descrição
+Este projeto implementa o **Assignment 1 – Diário de Bordo**,  utilizando a arquitetura Medalion Bronze → Silver → Gold em **PySpark**, com execução via **notebooks Databricks**. Cada camada é acionada como **task individual**, promovendo modularidade, isolamento e facilidade de manutenção.
 
-Este projeto processa dados de corridas de um aplicativo de transporte privado, gerando uma tabela agregada **`rides_daily_info_gold`** com métricas diárias de uso, organizada em camadas seguindo a arquitetura **Medallion (Bronze → Silver → Gold)**.
+* **Bronze**: ingestão contínua via Auto Loader → tabela Delta bronze (`rides_bronze`)
+* **Silver**: parsing de datas, cast, validações → tabela Delta silver (`rides_silver`)
+* **Gold**: agregações diárias com métricas específicas → tabela Delta gold (`rides_daily_info_gold`)
 
-A solução foi implementada inteiramente no **Databricks** utilizando **Databricks SQL** e **Streaming Tables**, permitindo ingestão contínua em caso de adição de novos arquivos.
+## Principais Componentes e Decisões de implementação
+
+* **Modularização com classes Python** (`src/stages.py`): cada estágio (BronzeToSilver, SilverToGold) é um `DataFrame -> DataFrame` puramente funcional facilitando o reuso.
+
+* **Constantes e utilitários centralizados** (`src/constants.py` e `src/schemas.py` ): esquema, padrões de data, paths, normalização de strings, mantendo consistência e fácil manutenção.
+
+
+## Estrutura do Repositório
+
+```
+assignment_1/
+├── notebooks/
+│   ├── bronze.ipynb
+│   ├── silver.ipynb
+│   └── gold.ipynb
+├── src/
+│   ├── stages.py
+│   ├── constants.py
+|   └── schemas.py
+├── README.md
+└── requirements.txt
+```
 
 ![ETL Job Success](img/1-ETL-job.png)
+*Figura 1: Imagem ilustrativa de execução bem-sucedida do job ETL no Databricks.*
 
 ## 🚀 Setup no Databricks
 
-1. **Carregar o dataset** para a pasta `/Volumes/sandbox_first_catalog/transports/landing-zone/` no **Unity Catalog**.
-2. Criar um **Job** no Databricks e associar o arquivo `job.sql`.
-3. Configurar para execução **Streaming**.
-4. Monitorar o job pela UI do Databricks.
-5. Consultar a tabela final com:
+1. No **Databricks**, suba o diretório `assignment_1/`, configure os notebooks como tasks de um job.
 
-   ```sql
-   SELECT * FROM sandbox_first_catalog.transports.rides_daily_info_gold;
-   ```
+2. Para simular a chegada do dado na landing, suba o csv para o local `/Volumes/sandbox_first_catalog/transports/landing-zone/`no volume DBFS, dentro do catalogo `sandbox_first_catalog` e do schema `transports`
 
+2. Execute sequencialmente: **01\_bronze → 02\_silver → 03\_gold** através de trigger manual
+
+3. Visualiza as tabelas no catálogo conforme tabela abaixo.
 
 ## ⚙️ Camadas do Pipeline
 
-1. **Bronze** — Ingestão dos arquivos CSV da landing zone (DBFS)
+| Camada     | Descrição                          | Catálogo                                   |
+| ---------- | ---------------------------------- | ----------------------------------------- |
+| **Bronze** | Ingestão raw CSV com Auto Loader   | `sandbox_first_catalog.transports.rides_bronze` |
+| **Silver** | Parsing, validações, transformação | `sandbox_first_catalog.transports.rides_silver` |
+| **Gold**   | Agregações diárias e métricas      | `sandbox_first_catalog.transports.rides_daily_info_gold` |
 
-   ```sql
-   CREATE OR REFRESH STREAMING TABLE sandbox_first_catalog.transports.rides_bronze
-   AS SELECT *
-   FROM STREAM read_files('/Volumes/sandbox_first_catalog/transports/landing-zone/', ...);
-   ```
-
-2. **Silver** — Limpeza, conversão de tipos e aplicação de constraints de qualidade
-
-   ```sql
-   CREATE OR REFRESH STREAMING TABLE sandbox_first_catalog.transports.rides_silver (
-     CONSTRAINT valid_start_date EXPECT (DATA_INICIO IS NOT NULL),
-     CONSTRAINT valid_distance EXPECT (DISTANCIA IS NOT NULL)
-   )
-   AS SELECT ... FROM STREAM sandbox_first_catalog.transports.rides_bronze;
-   ```
-
-3. **Gold** — Agregações e métricas finais por dia como requisitado pelos consumidores
-
-   ```sql
-   CREATE OR REFRESH STREAMING TABLE sandbox_first_catalog.transports.rides_daily_info_gold
-   AS SELECT DATA_INICIO AS DT_REFE, COUNT(*) AS QT_CORR, ...
-   FROM STREAM sandbox_first_catalog.transports.rides_silver
-   GROUP BY DT_REFE;
-   ```
 
 ## 📊 Colunas de Saída (`rides_daily_info_gold`)
