+++ assignment_1/rides_ingestion/notebooks/bronze.py
@@ -1,8 +1,8 @@
 from pyspark.sql import DataFrame, SparkSession
 
-from assignment_1.rides_ingestion.src.base import Stage
 from assignment_1.rides_ingestion.src.constants import LANDING_PATH, TBL_BRONZE
 from assignment_1.rides_ingestion.src.schemas import BRONZE_SCHEMA
+from assignment_1.rides_ingestion.src.stage import Stage
 
 spark = SparkSession.builder.getOrCreate()
