+++ assignment_1/rides_ingestion/src/stage.py
